requirements.txt
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
idna==2.8
lxml==4.4.1
pycparser==2.19
pyOpenSSL==19.0.0
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
urllib3==1.25.6
user-agent==0.1.9
